requirements.txt
Flask==0.12
pytest==3.0.6
Flask-PyMongo
